Updated, more secured (safer).

+++ automap.sh
@@ -1,35 +1,41 @@
 #!/bin/sh
 
-USAGE="$0 <web root> : generates an html map from the available pages in <web root>"
+MAP_FILE="Map.html"
+
+USAGE="$0 <web root> : generates an html map from the available pages in <web root>, and generates a $MAP_FILE file in the current directory."
 
 WEB_ROOT=$1
 
 echo
 
-if [ -z "$WEB_ROOT" ]; then
-	echo "$USAGE : error, not enough parameters"
-	exit
+if [ -f "$MAP_FILE" ] ; then
+	echo "Error, a $MAP_FILE file already exists, remove it first."
+	exit 1
 fi
 
-if [ ! -d "$WEB_ROOT" ]; then
+if [ -z "$WEB_ROOT" ] ; then
+	echo "$USAGE : error, not enough parameters"
+	exit 2
+fi
+
+if [ ! -d "$WEB_ROOT" ] ; then
 	echo "$USAGE : error, $1 is not a directory"
-	exit
+	exit 3
 fi
 
  
-MAP_FILE="Map.html"
 MAP_HEADER="$WEB_ROOT/../common/Map-header.html"
 MAP_FOOTER="$WEB_ROOT/../common/Map-footer.html"
 
 
 if [ ! -f "$MAP_HEADER" ] ; then
 	echo "$USAGE : error for map header, file <$MAP_HEADER> does not exist"
-	exit
+	exit 4
 fi
 
 if [ ! -f "$MAP_FOOTER" ] ; then
 	echo "$USAGE : error for map header, file <$MAP_FOOTER> does not exist"
-	exit
+	exit 5
 fi
 
 	
@@ -37,7 +43,7 @@ echo "Generating map file $MAP_FILE from $WEB_ROOT"
 
 cat $MAP_HEADER > $MAP_FILE
 
-TARGET_FILES=`find $WEB_ROOT -name '*.html' -print`
+TARGET_FILES=`find $WEB_ROOT -name '*.html' -print | grep -v index.htm | grep -v Menu`
 
 echo "<ul>" >> $MAP_FILE
 
@@ -50,4 +56,4 @@ echo "</ul>" >> $MAP_FILE
 
 cat $MAP_FOOTER >> $MAP_FILE
 
-echo "Map generated !"
+echo "Map generated ! ($MAP_FILE)"
